+++ History.md
@@ -19,6 +19,7 @@
 - When exporting the proposed procedure list into ODS, there is now an option to also include the comments.
 - Bugfix: When a motion type was created from the scratch and motion sections positioned right were added, the layout didn't switch to two-column mode.
 - Bugfix: When an optional image was not uploaded, the LaTeX-based PDF export did not work.
+- Bugfix: It was not possible to delete admin comments in the proposed procedure.
 - Internal: translatable strings can now include a comment / description which will be displayed in the translation page.
 
 ### Version 4.0.2 (2018-09-15)